+++ luckydonaldUtils/djangos/__init__.py
@@ -1,6 +1,2 @@
 # -*- coding: utf-8 -*-
-from ..dependencies import import_or_install
-
 __author__ = 'luckydonald'
-
-import_or_install("django", "django")