+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Changelog
 =========
+0.48
+----
+- Improvements in `tree`, and `dependencies`.
+- Added `text.split_in_parts(string, parts, strict=False)`: Splits a string in given `parts` pieces.
+
 0.47
 ----
 - Restructured `files.py` into a cleaner structured module.
@@ -16,6 +21,7 @@ Changelog
         - `do_a_filename`
     - `files.temp`
         - `gettempdir`
+    - 
 - Newly added `files.tree.tree(directory, padding="", print_files=False, level=-1, print_it=True)`. 
     It comes executable from command line. Just call `python -m luckydonaldUtils.files.tree` 
 - Improved `functions.deprecated` to now also accept a message: